REPORTING_TOOL/Pytest_reports/requirements.txt
pytest==7.2.0
requests==2.28.1
Appium-Python-Client==2.7.1
reportportal-client
pytest-reportportal
pytest-html==3.2.0
allure-pytest
